+++ src/code-blocks/types.ts
@@ -1,13 +1,14 @@
 export type { ChoiceGroup, ChoiceGroupWithParent, ParentChoiceGroup }
 
-import type { Config } from '../types/Config.js'
+import type { Config, ChoiceItem } from '../types/Config.js'
 
-type ChoiceGroup = NonNullable<Config['choices']>[string] & {
+type ChoiceGroup = Omit<NonNullable<Config['choices']>[string], 'choices'> & {
   name: string
+  choices: ChoiceItem[]
   emptyChoices: string[]
   hidden: boolean
   lvl: number
   isBuiltIn?: boolean
 }
-type ParentChoiceGroup = { name: string; default: string }
+type ParentChoiceGroup = { name: string; default: string; emptyChoices: string[] }
 type ChoiceGroupWithParent = ChoiceGroup & { parentChoiceGroup?: ParentChoiceGroup & { choices: string[] } }